Moved new commander integration to former cli index
