+++ packages/cli/package.json
@@ -18,7 +18,7 @@
   },
   "dependencies": {
     "@medplum/core": "*",
-    "commander": "^10.0.0",
+    "commander": "^10.0.1",
     "dotenv": "16.0.3",
     "node-fetch": "2.6.9"
   },
