+++ packages/cli/src/auth.ts
