+++ packages/cli/src/bots.ts
