+++ packages/cli/src/legacy-bot.ts
